# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/student/forgot-password": {
      "post": {
        "operationId": "StudentController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotStudentPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Student"]
      }
    },
    "/api/v1/student/profile/me": {
      "put": {
        "operationId": "StudentController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Student"]
      }
    },
...
    "/api/v1/student/{id}/parents": {
      "get": {
        "operationId": "StudentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Student"]
      },
      "patch": {
        "operationId": "StudentController_updateParent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateParentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Student"]
      },
      "delete": {
        "operationId": "StudentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Student"]
      }
    },
    "/api/v1/student/{studentId}/parents": {
      "post": {
        "operationId": "StudentController_createParent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Student"]
      }
    },
...
    "/api/v1/student/classes/{classLevelId}/calendars/{calendarId}/attendance/grouped": {
      "get": {
        "operationId": "StudentController_getMyAttendanceGroupedByTermAndMonth",
        "parameters": [
          {
            "name": "classLevelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "calendarId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Student"]
      }
    },
...
    "/api/v1/student/assignments": {
      "get": {
        "operationId": "StudentController_getMyAssignments",
        "parameters": [
          {
            "name": "pending",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "submitted",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "graded",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Student"]
      }
    },
    "/api/v1/student/assignments/{id}/submit": {
      "post": {
        "operationId": "StudentController_submitAssignment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Student"]
      }
    },
...
  },
  "info": {
    "title": "School Management API",
    "description": "OpenAPI spec for Postman and other clients",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "otherName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "phoneContact": {
            "type": "string"
          },
          "PlaceOfBirth": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female"]
          },
          "DateOfBirth": {
            "type": "string"
          },
          "BoxAddress": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "streetAddress": {
            "type": "string"
          },
          "optionalPhoneContact": {
            "type": "string"
          },
          "optionalPhoneContactTwo": {
            "type": "string"
          }
        }
      },
...
      "ForgotStudentPasswordDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "example": "student@school.com or STU-2024-001"
          }
        },
        "required": ["identifier"]
      },
      "CreateParentDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "occupation": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "parent@example.com"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "relationship": {
            "type": "string"
          }
        },
        "required": ["firstName", "lastName"]
      },
      "UpdateParentDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "occupation": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "parent@example.com"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "relationship": {
            "type": "string"
          }
        }
      },
      "SubmitAssignmentDto": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string"
          }
        }
      },
...
    }
  }
}